Refactor auth services, leave only login and logout
