+++ src/main/java/lt/vu/trip/config/RestExceptionHandler.java
@@ -1,7 +1,7 @@
 package lt.vu.trip.config;
 
 import lombok.extern.slf4j.Slf4j;
-import lt.vu.trip.service.auth.InvalidJwtAuthenticationException;
+import lt.vu.trip.service.auth.jwt.InvalidJwtAuthenticationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
